[Refactor] DataStore 모듈로 따로 분리
