+++ app/build.gradle.kts
@@ -62,6 +62,7 @@ dependencies {
     implementation(project(":core:domain"))
     implementation(project(":core:navigator"))
     implementation(project(":core:remote"))
+    implementation(project(":core:datastore"))
     implementation(project(":feature:sample"))
     implementation(project(":feature:onboard"))
     implementation(project(":feature:detail"))
