+++ core/data/build.gradle.kts
@@ -22,7 +22,7 @@ dependencies {
     implementation(project(":core:domain"))
     implementation(project(":core:remote"))
     implementation(project(":core:local"))
-
+    implementation(project(":core:datastore"))
     implementation(libs.kotlin.coroutines.android)
     implementation(libs.kotlin.coroutines.core)
     implementation(libs.hilt.android)
