+++ core/local/src/main/java/team/ppac/local/datasource/UserLocalDataSource.kt
@@ -1,7 +1,7 @@
 package team.ppac.local.datasource
 
 import kotlinx.coroutines.flow.Flow
-import team.ppac.local.entity.UserData
+import team.ppac.datastore.entity.UserData
 
 interface UserLocalDataSource {
     val userDataFlow: Flow<UserData>
