+++ core/local/src/main/java/team/ppac/local/datasource/impl/UserDataSourceImpl.kt
@@ -3,7 +3,7 @@ package team.ppac.local.datasource.impl
 import androidx.datastore.core.DataStore
 import kotlinx.coroutines.flow.Flow
 import team.ppac.local.datasource.UserLocalDataSource
-import team.ppac.local.entity.UserData
+import team.ppac.datastore.entity.UserData
 import javax.inject.Inject
 
 internal class UserDataSourceImpl @Inject constructor(
