+++ settings.gradle.kts
@@ -38,3 +38,4 @@ include(":feature:mypage")
 include(":feature:detail")
 include(":feature:setting")
 include(":feature:splash")
+include(":core:datastore")
